Splitting start_pipelines
